refactor(file-upload-item): typo
