+++ packages/shared/src/human-file-size/index.ts
@@ -1,0 +1,1 @@
+export { humanFileSize } from './human-file-size';
